requirements.txt
pyjwt
websockets
requests
